Refactor: Move Algolia search index functionality to a new utility module and update related imports
